got axios working for more efficient fetches
