+++ app-frontend/package.json
@@ -10,6 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "axios": "^1.13.6",
     "bootstrap": "^5.3.8",
     "dotenv": "^17.3.1",
     "react": "^19.2.0",
