+++ app-frontend/src/components/helper-routes/ProtectedRoute.jsx
@@ -1,6 +1,6 @@
 //This is used to renavigate those who do not bear a token
 import { Navigate } from 'react-router-dom';
-import { isAuthenticated } from '../../services/api/auth';
+import { isAuthenticated } from '../../services/auth';
 
 export default function ProtectedRoute({ children }) {
     if(!isAuthenticated()){
