+++ app-frontend/src/services/api/api.js
@@ -1,0 +1,7 @@
+import axios from 'axios';
+
+const api = axios.create({
+    baseURL: import.meta.env.VITE_API_BASE_URL
+});
+
+export default api;
